de-muiify-evidence: Add EVResourceTypeIcon to EVResourceItemDialog and EVResourceItemTitleBar
