+++ src/app-web/components/EVResourceItemDialog.jsx
@@ -15,25 +15,8 @@ import './MEMEStyles.css';
 import './EVResourceItemDialog.css';
 
 import EVResourceItemTitleBar from './EVResourceItemTitleBar';
+import EVResourceTypeIcon from './EVResourceTypeIcon';
 import ICNCountBadge from './ICNCountBadge';
-
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faImage } from '@fortawesome/free-solid-svg-icons';
-import { faFileLines } from '@fortawesome/free-solid-svg-icons';
-import { faLightbulb } from '@fortawesome/free-solid-svg-icons';
-import { faCircleQuestion } from '@fortawesome/free-solid-svg-icons';
-const ImageIcon = <FontAwesomeIcon icon={faImage} />;
-const DescriptionIcon = <FontAwesomeIcon icon={faFileLines} />;
-const IdeaIcon = <FontAwesomeIcon icon={faLightbulb} />;
-const ContactSupportIcon = <FontAwesomeIcon icon={faCircleQuestion} />;
-const RESOURCE_TYPES = {
-  simulation: ImageIcon,
-  assumption: IdeaIcon,
-  idea: IdeaIcon,
-  report: DescriptionIcon,
-  question: ContactSupportIcon,
-  other: DescriptionIcon
-};
 
 /// COMPONENTS ////////////////////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
@@ -223,9 +206,7 @@ class EVResourceItemDialog extends React.Component {
             <div>Type:&nbsp;</div>
             <div>
               {resource.type}&nbsp;
-              {RESOURCE_TYPES[resource.type]
-                ? RESOURCE_TYPES[resource.type]
-                : RESOURCE_TYPES.other}
+              <EVResourceTypeIcon type={resource.type} />
             </div>
           </div>
           <div className="box">
